Calculate ball heading, UDP, proper red hue
The red hue was not able to be picked up properly because it existed at both
ends of the hue spectrum. This code now puts red at the center by
converting from BGR to RGB before converting to HSV.

There was also some work on getting UDP to working. Not tested :)

Heading of the ball while it is moving is now calculated.