Fix template functions, flood of raw_img files
The flood of raw_img files was caused by the dirname str being set to . in the
demo() function after it had already been declared and directory created in the
main() function.

The program failed to compile when ball tracking was enabled due to the use of
the template function square inside of the code for the ball tracking because
apparently template functions implementations are supposed to be declared in the
header file instead of in source files. Odd. Note: Ask sanjiv/cezary. Todo: Move
the rest of the already existing template functions into the header file.

Also, OpenCV now works properly on mac! :D