Filter out bumpers from image
The L-detection logic has been applied to ball tracking to filter out based on
a ratio. The ratio is:
  Contour Area
 -------------- = Circle Ratio
  Circle  Area

Values from simple initial testing measured out to be 0.6 on the minimum and
0.9 on the maximum, with average around 0.8. Keep in mind these values are not
from images that include bumpers; only from samples on carpet in LHS cafeteria.

The circle area is always larger than the contour area (unless
minEnclosingCircle changes it's implementation). The circle area is expected to
be much larger than the contour area if the object is not a ball. A ball would
show up mostly as a circle hopefully, where a bumper would be a long rectangle.
The minEnclosingCircle would find the entire length of the rectangle as it's
radius and not much of the red contour would show up as it would only include
the bumper itself.