Add depth ball tracking support
The program now gathers data from the ASUS Xtion camera, logs it (at least), and
we attempt to track balls. We have not tested the ball tracking yet but we are
adding it as a possibility for the future. The program will log data it receives
from the ball tracking running on top of the depth code and the images. This
will allow us to gather data from the competition this year that we can apply to
competitions in the future.

We have also made a camera mount for the ASUS Xtion and printed several
prototypes. This will be mounted on the robot with an extra ODROID. It will be
used to log information in the match that we can study for potential use in
future competitions. We do not currently know the effectiveness of depth in a
live competition.

I have also worked on this support in a few new classes, they way I would like
to implement all the input styles in the program when I work on optimizing /
cleaning in the offseason.