(Demo) "IMGDIR" loading support
The program now supports a new image backend called "IMGDIR". This
option loads images in sequence from a folder instead of from a
camera or just a single image. This can be used as a method of post-
processing that runs the actual program instaed of simply loading
contours from a log file and adding them to the screen. More
information about what actually happened from the robot can be
viewed, such as distance to targets and whether one was an "L" or
not.